+++ lib/src/list_config/loading_more_list_config.dart
@@ -5,8 +5,19 @@ import 'package:waterfall_flow/waterfall_flow.dart';
 import '../indicator_widget.dart';
 import 'sliver_list_config.dart';
 
+typedef LoadingMoreItemBuilder<T> = Widget Function(
+  BuildContext context,
+  T item,
+  int index,
+);
+
+typedef LoadingMoreIndicatorBuilder = Widget? Function(
+  BuildContext context,
+  IndicatorStatus status,
+);
+
 class LoadingMoreListConfig<T> {
-  LoadingMoreListConfig(
+  const LoadingMoreListConfig(
     this.itemBuilder,
     this.sourceList, {
     this.indicatorBuilder,
@@ -19,31 +30,38 @@ class LoadingMoreListConfig<T> {
     this.childCountBuilder,
     this.getActualIndex,
   });
-  //Item builder
-  final Widget Function(BuildContext context, T item, int index) itemBuilder;
 
-  //source list
+  /// Builds widget from [item] and [index].
+  final LoadingMoreItemBuilder<T> itemBuilder;
+
+  /// Source list based on the [LoadingMoreBase].
   final LoadingMoreBase<T> sourceList;
 
-  //widget builder for different loading state
-  //the deafault is LoadingIndicatorWidget
+  /// The builder for each [IndicatorStatus].
+  ///
+  /// You can handle cases that you need to customize, and return null result
+  /// to fallback to the default indicator.
+  ///
+  /// Defaults to [IndicatorWidget].
   final LoadingMoreIndicatorBuilder? indicatorBuilder;
 
-  //whether is gird view
+  /// Set the [gridDelegate] if the config is for a grid view.
   final SliverGridDelegate? gridDelegate;
 
-  //whether auto call sourceList.loadmore when meet the condition
-  //if false, you must call sourceList.loadmore by yourself.
+  /// Whether to auto load more with the [sourceList].
+  ///
+  /// Set to false if you want to handle the load more action yourself.
   final bool autoLoadMore;
 
-  /// The delegate for WaterfallFlow or ExtendedList.
+  /// The delegate for [WaterfallFlow] or [ExtendedList].
   final ExtendedListDelegate? extendedListDelegate;
 
-  /// Layout type of last child
+  /// Layout type of last child.
   final LastChildLayoutType lastChildLayoutType;
 
-  //whether auto call sourceList.refresh when first load
-  //if false, you must call sourceList.refresh by yourself.
+  /// Whether to auto refresh with the [sourceList].
+  ///
+  /// Set to false if you want to handle the refresh action yourself.
   final bool autoRefresh;
 
   /// The total number of children this delegate can provide.
@@ -52,14 +70,18 @@ class LoadingMoreListConfig<T> {
   /// [builder] returns null.
   final int? childCount;
 
-  /// The builder to get child count,the input is sourceList.length
+  /// The builder to get child count,the input is [LoadingMoreBase.length].
   final int Function(int count)? childCountBuilder;
 
   final int Function(int int)? getActualIndex;
 
-  bool get isSliver {
-    return this is SliverListConfig<T>;
-  }
+  bool get isSliver => this is SliverListConfig<T>;
+
+  bool get hasMore => sourceList.hasMore;
+
+  bool get hasError => sourceList.hasError;
+
+  bool get isLoading => sourceList.isLoading;
 
   Widget? buildContent(BuildContext context, LoadingMoreBase<T>? source) {
     //from stream builder or from refresh
@@ -74,30 +96,23 @@ class LoadingMoreListConfig<T> {
                 this as SliverListConfig<dynamic>;
             // prevent lock list load
             if (sliverListConfig.actualLock) {
-              return SliverToBoxAdapter(
-                child: Container(),
-              );
-            } else {
-              sourceList.refresh();
+              return const SliverToBoxAdapter(child: SizedBox.shrink());
             }
-          } else {
-            sourceList.refresh();
           }
+          sourceList.refresh();
         }
       }
       Widget? widget;
-      if (indicatorBuilder != null)
+      if (indicatorBuilder != null) {
         widget = indicatorBuilder!(context, IndicatorStatus.fullScreenBusying);
+      }
       widget = widget ??
           IndicatorWidget(
             IndicatorStatus.fullScreenBusying,
             isSliver: isSliver,
           );
-
       return widget;
-    }
-    //empty
-    else if (source.isEmpty &&
+    } else if (source.isEmpty &&
         (source.indicatorStatus == IndicatorStatus.empty ||
             source.indicatorStatus == IndicatorStatus.fullScreenError)) {
       Widget? widget1;
@@ -108,17 +123,11 @@ class LoadingMoreListConfig<T> {
             sourceList.indicatorStatus,
             isSliver: isSliver,
             tryAgain: source.indicatorStatus == IndicatorStatus.fullScreenError
-                ? () {
-                    sourceList.errorRefresh();
-                  }
+                ? sourceList.errorRefresh
                 : null,
           );
       return widget1;
     }
-    //show list
-    //else if (source.length > 0) {
-
-    // }
     return null;
   }
 
@@ -144,11 +153,7 @@ class LoadingMoreListConfig<T> {
       if (indicatorBuilder != null) {
         widget1 = indicatorBuilder!(context, status);
       }
-      widget1 = widget1 ??
-          IndicatorWidget(
-            status,
-            isSliver: isSliver,
-          );
+      widget1 = widget1 ?? IndicatorWidget(status, isSliver: isSliver);
       return widget1;
     }
     return itemBuilder(
@@ -165,21 +170,20 @@ class LoadingMoreListConfig<T> {
       if (indicatorBuilder != null)
         widget = indicatorBuilder!(context, IndicatorStatus.error);
       widget = widget ??
-          IndicatorWidget(IndicatorStatus.error, isSliver: isSliver,
-              tryAgain: () {
-            sourceList.errorRefresh();
-          });
+          IndicatorWidget(
+            IndicatorStatus.error,
+            isSliver: isSliver,
+            tryAgain: sourceList.errorRefresh,
+          );
       return widget;
     }
     return null;
   }
 
-  bool get hasMore => sourceList.hasMore;
-  bool get hasError => sourceList.hasError;
-  bool get isLoading => sourceList.isLoading;
-
-  ExtendedListDelegate getExtendedListDelegate(int childCount,
-      {bool showNoMore = true}) {
+  ExtendedListDelegate getExtendedListDelegate(
+    int childCount, {
+    bool showNoMore = true,
+  }) {
     if (extendedListDelegate != null) {
       if (extendedListDelegate
           is SliverWaterfallFlowDelegateWithFixedCrossAxisCount) {
@@ -236,13 +240,10 @@ class LoadingMoreListConfig<T> {
 
     return ExtendedListDelegate(
       lastChildLayoutTypeBuilder: showNoMore
-          ? ((int index) => childCount == index
+          ? (int index) => childCount == index
               ? lastChildLayoutType
-              : LastChildLayoutType.none)
+              : LastChildLayoutType.none
           : null,
     );
   }
 }
-
-typedef LoadingMoreIndicatorBuilder = Widget Function(
-    BuildContext context, IndicatorStatus status);